directory structure cleanup2

+++ src/TrainBuilder.py
@@ -332,7 +332,8 @@ class TrainBuilder:
         if not os.path.isdir(self.build_dir):
             os.mkdir(self.build_dir)
         else:
-            shutil.rmtree(os.path.join(self.build_dir, "pht_train"))
+            if os.path.isdir(os.path.join(self.build_dir, "pht_train")):
+                shutil.rmtree(os.path.join(self.build_dir, "pht_train"))
         base_path = os.path.join(self.build_dir, "pht_train")
         os.mkdir(base_path)
         ep_dir = os.path.join(base_path, 'endpoints')
